AP: InstallWinget: check for unsupported build before installing

+++ CW-CLI.ps1
@@ -305,23 +305,31 @@ Function UnpinStartTiles {
 # Install WinGet
 Function InstallWinGet {
 	Write-Host " "
-	# Import BitsTransfer module and download NetTestFile
-    Import-Module BitsTransfer 
-    Start-BitsTransfer https://raw.githubusercontent.com/CleanWin/Files/main/NetTestFile
-    # If the file exists, proceed with downloading WinGet files, else inform user about no internet connection.
-    If (Test-Path NetTestFile) {
-        Remove-Item NetTestFile
-        Write-Host "Downloading WinGet..."
-        Start-BitsTransfer https://github.com/CleanWin/Files/raw/main/Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.appxbundle
-        Start-BitsTransfer https://github.com/CleanWin/Files/raw/main/Microsoft.VCLibs.140.00.UWPDesktop_14.0.29231.0_x64__8wekyb3d8bbwe.Appx
-        Write-Host "Installing WinGet..."
-        Add-AppxPackage -Path .\Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.appxbundle -DependencyPath .\Microsoft.VCLibs.140.00.UWPDesktop_14.0.29231.0_x64__8wekyb3d8bbwe.Appx
-        Remove-Item Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.appxbundle
-        Remove-Item Microsoft.VCLibs.140.00.UWPDesktop_14.0.29231.0_x64__8wekyb3d8bbwe.Appx
-        Write-Host "Done."
-	} 
-	else {
-	  Write-Host "We can't connect to GitHub to download the installation files. Are you sure that your internet connection is working?"
+	If ([System.Environment]::OSVersion.Version.Build -ge 16299) {
+		# Import BitsTransfer module and download NetTestFile
+		Import-Module BitsTransfer 
+		Start-BitsTransfer https://raw.githubusercontent.com/CleanWin/Files/main/NetTestFile
+			# If the file exists, proceed with downloading WinGet files, else inform user about no internet connection.
+			If (Test-Path NetTestFile) {
+				Remove-Item NetTestFile
+				Write-Host "Downloading WinGet..."
+				Start-BitsTransfer https://github.com/CleanWin/Files/raw/main/Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.appxbundle
+				Start-BitsTransfer https://github.com/CleanWin/Files/raw/main/Microsoft.VCLibs.140.00.UWPDesktop_14.0.29231.0_x64__8wekyb3d8bbwe.Appx
+				Write-Host "Installing WinGet..."
+				Add-AppxPackage -Path .\Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.appxbundle -DependencyPath .\Microsoft.VCLibs.140.00.UWPDesktop_14.0.29231.0_x64__8wekyb3d8bbwe.Appx
+				Remove-Item Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.appxbundle
+				Remove-Item Microsoft.VCLibs.140.00.UWPDesktop_14.0.29231.0_x64__8wekyb3d8bbwe.Appx
+				Write-Host "Done."
+			} 
+			else {
+			Write-Host "We can't connect to GitHub to download the installation files. Are you sure that your internet connection is working?"
+			}
+	}
+	Else {
+		$version = Get-ItemPropertyValue "HKLM:\SOFTWARE\Microsoft\Windows NT\CurrentVersion" -Name DisplayVersion
+		$osbuild = Get-ItemPropertyValue "HKLM:\SOFTWARE\Microsoft\Windows NT\CurrentVersion" -Name CurrentBuild
+		Write-Host "WinGet can't be installed since you're running Windows 10 $version build $osbuild which is unsupported."
+		Write-Host "Its strongly recommended that you upgrade to or fresh install the latest version of Windows 10."
 	}
 }
